[fix][change]更改单更新时间传递&已执行指令不可变更

+++ wip-domain/src/main/java/com/cvte/scm/wip/domain/core/requirement/entity/ReqInsEntity.java
@@ -133,6 +133,11 @@ public class ReqInsEntity implements Entity<String> {
      * @author xueyuting
      */
     public ReqInsEntity completeInstruction(ReqInsBuildVO vo) {
+        ReqInsEntity existsIns = this.getByKey(vo.getInsHeaderId());
+        if (Objects.nonNull(existsIns) && !ProcessingStatusEnum.PENDING.getCode().equals(existsIns.getStatus())) {
+            log.info("指令(ID={})已执行,不可变更", vo.getInsHeaderId());
+            return null;
+        }
         if (ChangedTypeEnum.DELETE.getCode().equals(vo.getChangeType())) {
             return this.deleteCompleteReqIns(vo);
         } else if (ChangedTypeEnum.UPDATE.getCode().equals(vo.getChangeType())) {
